Added AsyncApply method to the DB type
This method just applies the given function asynchronously
